pyproject.toml
[project]
name = "bangumi-syncer"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.126.0",
    "ijson>=3.4.0.post0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.38.0",
    "httpx>=0.25.0",
    "APScheduler>=3.10.0",
]

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "mypy>=1.19.1",
    "ruff>=0.14.10",
    "ty>=0.0.9",
    "playwright>=1.57.0",
    "pytest-playwright>=0.4.0",
    "pytest-asyncio>=1.2.0",
    "pytest-bdd>=7.0.0",
    "pytest-cov>=6.0.0",
    "responses>=0.25.0",
    "respx>=0.22.0",
]

[tool.ruff]
# 设定目标 Python 版本
target-version = "py39"
line-length = 88

[tool.ruff.lint]
# 建议开启的初始规则集
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes (错误检测)
    "I",   # isort (自动整理 import)
    "UP",  # pyupgrade (自动升级旧语法)
    "B",   # flake8-bugbear (发现潜在设计缺陷)
]
ignore = [
    "E501",  # 行过长（由 ruff format 处理）
    "E722",  # bare except
    "B008",  # 函数调用作为默认参数
    "B904",  # raise 不带 from（简化异常处理）
]

[tool.ruff.lint.isort]
# 自动合并同类 import
combine-as-imports = true

[tool.djlint]
# 模板引擎类型
profile = "jinja"

# 缩进空格数
indent = 4

# 最大行长度
max_line_length = 120

[tool.pytest.ini_options]
# 基础配置
minversion = "6.0"
pythonpath = "."    # ✅ 关键修复：将根目录加入搜索路径
addopts = "-ra -q --durations=5 --feature tests/features"  # 运行时参数：显示额外汇总、精简输出、列出最慢的5个测试
testpaths = [
]

# 匹配规则
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*", "Check*"]
python_functions = ["test_*", "check_*"]

# pytest-bdd 配置
bdd_features_base_dir = "tests/features"

# 环境变量或日志配置（可选）
log_cli = true
log_cli_level = "INFO"

asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"  # 消除新版 pytest-asyncio 的 Scope 警告

# ✅ 新增：注册自定义标记，消除 PytestUnknownMarkWarning
markers = [
    "asyncio: mark test as async",
    "performance: mark test as a performance test",
    "playwright: mark test as a Playwright E2E test",
]

# Playwright 配置
[tool.playwright]
# 浏览器配置
browser = "chromium"
# 是否无头模式运行
headless = true
# 是否捕获视频
video = "retain-on-failure"
# 是否捕获跟踪
trace = "retain-on-failure"
# 截图目录
screenshot = "only-on-failure"

[tool.playwright.options]
# 浏览器启动选项
launch_options = { args = ["--disable-dev-shm-usage", "--no-sandbox"] }
# 视口大小
viewport = { width = 1280, height = 720 }
# 区域设置
locale = "zh-CN"
timezone_id = "Asia/Shanghai"

# Coverage configuration
[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
show_missing = true
skip_empty = true
